# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: FETAL-GUARD API
  description: |
    Backend API for FETAL-GUARD: Wearable Early Warning System for fetal monitoring.
    Provides endpoints for patient management, device pairing, monitoring sessions,
    alerts, and clinician dashboard.
  version: 1.0.0
  contact:
    name: FETAL-GUARD Team
    email: support@fetal-guard.id
  license:
    name: Proprietary
servers:
  - url: https://api.fetal-guard.id/v1
    description: Production server
  - url: https://staging-api.fetal-guard.id/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: List all patients (clinician only)
      parameters:
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /sessions:
    get:
      tags: [Sessions]
      summary: List monitoring sessions
      parameters:
        - name: patient_id
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Sessions]
      summary: Start a new monitoring session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
...
  /clinician/patients:
    get:
      tags: [Clinician]
      summary: Get clinician's patient list
      parameters:
        - name: risk_filter
          in: query
          schema:
            type: string
            enum: [all, high, active_alert]
      responses:
        '200':
          description: Patient list for clinician
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClinicianPatientView'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    PatientSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        gestational_age_weeks:
          type: integer
        last_session_date:
          type: string
          format: date-time
        current_risk_level:
          type: string
          enum: [low, medium, high]
        active_alerts:
          type: integer
...
    Session:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        device_id:
          type: string
        device_serial:
          type: string
        device_model:
          type: string
        device_firmware_version:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        duration_seconds:
          type: integer
        time_zone:
          type: string
          example: "Asia/Jakarta"
        sampling_rate_hz:
          type: integer
          example: 250
        channel_count:
          type: integer
          example: 4
        status:
          type: string
          enum: [active, stopped, completed, error]
        fhr_summary:
          type: object
          properties:
            baseline_bpm:
              type: number
            min_bpm:
              type: number
            max_bpm:
              type: number
            mean_bpm:
              type: number
            variability:
              type: number
        events:
          type: array
          items:
            $ref: '#/components/schemas/SessionEvent'
        risk_assessment:
          $ref: '#/components/schemas/RiskAssessment'
        signal_quality:
          type: object
          properties:
            overall_percentage:
              type: number
            snr:
              type: number
            motion_ratio:
              type: number
            maternal_overlap_flag:
              type: boolean
        encryption_metadata:
          type: object
          properties:
            algorithm:
              type: string
              example: "AES-256-GCM"
            key_id:
              type: string
        notes:
          type: string
        clinician_notes:
          type: string
    SessionSummary:
      type: object
      properties:
        session_id:
          type: string
        start_time:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        mean_fhr:
          type: number
        risk_level:
          type: string
          enum: [low, medium, high]
        event_count:
          type: integer
...
    SessionEvent:
      type: object
      properties:
        event_id:
          type: string
        type:
          type: string
          enum: [acceleration, deceleration, movement, maternal_signal, artifact]
        timestamp:
          type: string
          format: date-time
        duration_ms:
          type: integer
        severity:
          type: string
          enum: [mild, moderate, severe]
        details:
          type: object
    RiskAssessment:
      type: object
      properties:
        score:
          type: number
          minimum: 0
          maximum: 100
        level:
          type: string
          enum: [low, medium, high]
        confidence:
          type: number
        confidence_interval:
          type: object
          properties:
            lower:
              type: number
            upper:
              type: number
        model_version:
          type: string
        explainability:
          type: string
        contributing_factors:
          type: array
          items:
            type: object
            properties:
              factor:
                type: string
              impact:
                type: number
...
    ClinicianPatientView:
      type: object
      properties:
        patient_id:
          type: string
        name:
          type: string
        gestational_age_weeks:
          type: integer
        last_session:
          type: string
          format: date-time
        current_risk:
          type: string
        active_alerts:
          type: integer
        last_fhr:
          type: number
...